pyproject.toml
[project]
name = "telegram-bot-download-video"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
    "pytubefix>=9.1.1",
]
